web/requirements.txt
Flask==2.2.5
Flask-MySQL==1.5.2
Flask-login
Jinja2==3.1.2
markupsafe==2.1.1
itsdangerous==2.0.1
werkzeug==3.0.1
cryptography
Pillow